# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BirthdayInputReminderDaysBefore = typeof BirthdayInputReminderDaysBefore[keyof typeof BirthdayInputReminderDaysBefore];


export const BirthdayInputReminderDaysBefore = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_7: 7,
} as const;

export interface BirthdayInput {
  name: string;
  nickname?: string;
  dateOfBirth: string;
  birthYearKnown: boolean;
  phone?: string;
  email?: string;
  relationship?: string;
  notes?: string;
  reminderEnabled: boolean;
  reminderDaysBefore: BirthdayInputReminderDaysBefore;
  reminderTime: string;
}

/**
 * Any subset of birthday fields
 */
export type BirthdayPatch = BirthdayInput;

export type Birthday = BirthdayInput & {
  id: string;
  createdAt: string;
  updatedAt: string;
  deletedAt?: string;
};

export interface BirthdaySummary {
  total: number;
  monthCounts: number[];
  storage: string;
}

export type ListBirthdaysParams = {
includeDeleted?: boolean;
};