# lib/api-zod/src/generated/types/birthday.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BirthdayInput } from './birthdayInput';

export type Birthday = BirthdayInput & {
  id: string;
  createdAt: string;
  updatedAt: string;
  deletedAt?: string;
};