# lib/api-zod/src/generated/types/birthdayInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BirthdayInputReminderDaysBefore } from './birthdayInputReminderDaysBefore';

export interface BirthdayInput {
  name: string;
  nickname?: string;
  dateOfBirth: string;
  birthYearKnown: boolean;
  phone?: string;
  email?: string;
  relationship?: string;
  notes?: string;
  reminderEnabled: boolean;
  reminderDaysBefore: BirthdayInputReminderDaysBefore;
  reminderTime: string;
}