# lib/api-zod/src/generated/types/birthdayPatch.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BirthdayInput } from './birthdayInput';

/**
 * Any subset of birthday fields
 */
export type BirthdayPatch = BirthdayInput;