# lib/api-zod/src/generated/types/birthdaySummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BirthdaySummary {
  total: number;
  monthCounts: number[];
  storage: string;
}